Cleanup

+++ Boolder/Area/Map/HeadingView.swift
@@ -25,7 +25,7 @@ class HeadingView: UIView {
     
     var headingAccuracy: CLLocationDirectionAccuracy = 90 {
         didSet {
-            headingArcLayer.path = arcPath(size: bounds.size, angleInDegrees: headingAccuracy)
+            headingArcLayer.path = arcPath(size: bounds.size, angleInDegrees: CGFloat(headingAccuracy))
             headingArcLayer.isHidden = (headingAccuracy >= 90)
         }
     }
@@ -55,13 +55,13 @@ class HeadingView: UIView {
         }
     }
     
-    private func arcPath(size: CGSize, angleInDegrees: Double) -> CGPath {
+    private func arcPath(size: CGSize, angleInDegrees: CGFloat) -> CGPath {
         
         let bezierPath = UIBezierPath(
             arcCenter: CGPoint(x: size.width/2, y: size.height/2),
             radius: size.height/2,
-            startAngle: CGFloat(-90-angleInDegrees) * CGFloat(Double.pi) / 180,
-            endAngle: CGFloat(-90+angleInDegrees) * CGFloat(Double.pi) / 180,
+            startAngle: (-90-angleInDegrees) * CGFloat.pi / 180,
+            endAngle: (-90+angleInDegrees) * CGFloat.pi / 180,
             clockwise: true
         )
